# docs/design/personnel-mdm-openapi.yaml
openapi: 3.1.0
info:
  title: Lampang PAO - Central Personnel Master Data API
  version: 1.0.0
  summary: Single Source of Truth for personnel data of Lampang Provincial Administrative Organization (อบจ.ลำปาง)
  description: |
    API กลางสำหรับข้อมูลข้าราชการและพนักงาน อบจ.ลำปาง (Central Personnel MDM)

    **หลักการที่ผู้เรียกใช้ต้องทราบ**
    - ระบบปลายทางอ้างอิงบุคลากรด้วย `personId` (UUID) เท่านั้น เลขบัตรประชาชน (pid) จะไม่ปรากฏใน response ใด ๆ
      ยกเว้น `POST /persons/{personId}/pid` ซึ่งต้องมี scope `personnel:read:pid` และต้องระบุเหตุผลทุกครั้ง
    - การเข้าถึงข้อมูลควบคุมแบบ field-level ด้วย OAuth2 scope จาก Keycloak
      แต่ละกลุ่มฟิลด์ระบุ scope ที่ต้องใช้ไว้ใน `x-required-scope` ของ schema ฟิลด์ที่ไม่มีสิทธิ์จะถูกตัดออกจาก response (ไม่ใช่ null)
    - บุคคลสถานะ `INACTIVE` (ลาออก/โอนย้าย) เห็นได้เฉพาะ token ที่มี scope `personnel:read:inactive` ทั้งใน `GET /persons` และ endpoint รายบุคคล
      (`GET /persons/{id}`, `/photo`, `/employment`, `/change-log`, `/manage-profile`, `/history`, `POST /persons/{id}/pid`) ไม่มี scope -> 403 `insufficient-scope`
      (ไม่เขียน access log) บุคคลที่ไม่มีอยู่ยังได้ 404
    - ทุกคำขอที่คืนข้อมูลส่วนบุคคลถูกบันทึกใน access log (PDPA audit trail) พร้อม `purpose` ของระบบผู้เรียก
    - ข้อมูลระบุตัวตน (ชื่อ-สกุล วันเกิด ที่อยู่ตามทะเบียนบ้าน รูปถ่าย) มาจาก ThaID เท่านั้น แก้ไขผ่าน API ไม่ได้
    - เหตุการณ์เปลี่ยนแปลงข้อมูลส่งถึงระบบปลายทางสองทาง: webhook (push, มีลายเซ็น HMAC) และ `GET /events` (pull ด้วย cursor)
...

...
  contact:
...
servers:
...
security:
  - keycloak: []
...
paths:
...
  /persons/{personId}/manage-profile:
    parameters:
      - $ref: '#/components/parameters/PersonId'
    get:
      tags: [Provisioning]
      summary: Profile for the HR edit forms (expected identity, contact, emergency contacts)
      description: |
        ข้อมูลสำหรับหน้าแก้ไขของ HR: ข้อมูลระบุตัวตนที่ HR กรอกไว้ (รอยืนยันด้วย ThaID) พร้อมสถานะแก้ได้/ไม่ได้, ข้อมูลติดต่อ (ค่าเต็ม) และผู้ติดต่อฉุกเฉิน
        ข้อมูลติดต่อส่วนตัวเห็นได้เฉพาะ endpoint นี้ภายใต้ role `hr_master_data_admin` - `hr_officer` ทั่วไปไม่เห็นทั้งที่นี่และใน `GET /persons/{personId}`
        (ตัดกลุ่ม contact/emergencyContacts แม้ token จะมี `personnel:read:contact`)
        **สิทธิ์:** scope `personnel:manage:person` **และ** realm role `hr_master_data_admin` (ตรวจที่ API ทุก request; ไม่มี role -> 403 `insufficient-role`)
        ทุกครั้งที่เปิดดูถูกบันทึกใน access log
      operationId: getManageProfile
      security:
        - keycloak: [personnel:manage:person]
      responses:
        '200':
          description: โปรไฟล์สำหรับฟอร์มแก้ไข
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ManageProfile' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /persons/{personId}/expected-identity:
    parameters:
      - $ref: '#/components/parameters/PersonId'
    patch:
      tags: [Provisioning]
      summary: Edit the identity data HR expects (names, birth date) of a person who never verified with ThaID
      description: |
        แก้ชื่อ-นามสกุลไทยและวันเกิดที่ HR กรอกไว้ (คอลัมน์ `expected_*` บน `mdm.person` ไม่แตะ `person_identity`) ส่งเฉพาะฟิลด์ที่แก้ `birthDate: null` ล้างค่า
        ThaID เป็นหลักเสมอ: ตอน claim ค่าจาก ThaID เขียนลง `person_identity` โดยไม่ดูค่าเหล่านี้ (ไม่ตรงกันก็เก็บเฉยๆ ไม่แจ้งเตือน)
        **คนที่เคยยืนยันแล้ว** (`thaid_verified_at` หรือ `claimed_at` ไม่เป็น NULL หรือสถานะไม่ใช่ `PENDING_CLAIM`) -> 409 `identity-locked` ที่ API และ trigger ใน DB
        กันซ้ำอีกชั้น (การแข่งกับ `/sync/thaid` ถูกกันด้วยการล็อกแถว person `FOR UPDATE` ทั้งสองฝั่ง)
        **สิทธิ์:** scope `personnel:manage:person` **และ** realm role `hr_master_data_admin` (ไม่มี role -> 403 `insufficient-role`)
        **ทุกการเขียน:** `expectedVersion` บังคับ (ไม่ตรง -> 409 `version-conflict`), `reason` บังคับ (ว่าง -> 422 `reason-required`, มีเลข 13 หลัก -> 422
        `reason-contains-pid`), บันทึก `data_change_log` พร้อมผู้กระทำ, บุคคลที่ยัง `PENDING_CLAIM` ไม่ส่ง outbox; คืน `ManageProfile` ล่าสุด
      operationId: updateExpectedIdentity
      security:
        - keycloak: [personnel:manage:person]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExpectedIdentityPatch' }
      responses:
        '200':
          description: แก้แล้ว (หรือไม่มีอะไรเปลี่ยน)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ManageProfile' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /persons/{personId}/contact:
    parameters:
      - $ref: '#/components/parameters/PersonId'
    patch:
      tags: [Provisioning]
      summary: Edit contact data (partial update)
      description: |
        แก้ข้อมูลติดต่อแบบ PATCH: ฟิลด์ที่ไม่ส่ง = ไม่แตะ, ส่ง `null` หรือสตริงว่าง = ล้างค่า (เจ้าของยังแก้เองผ่าน `PUT /me/contact` ได้เหมือนเดิม)
        ค่าเต็มของฟิลด์ที่เปลี่ยนถูกเก็บใน `data_change_log` (หน้า DPO/ประวัติปกปิดค่า)
        **สิทธิ์:** scope `personnel:manage:person` **และ** realm role `hr_master_data_admin` (ไม่มี role -> 403 `insufficient-role`)
        **ทุกการเขียน:** `expectedVersion` บังคับ (ไม่ตรง -> 409 `version-conflict`), `reason` บังคับ (ว่าง -> 422 `reason-required`, มีเลข 13 หลัก -> 422
        `reason-contains-pid`), บันทึก `data_change_log` พร้อมผู้กระทำ, บุคคลที่ยัง `PENDING_CLAIM` ไม่ส่ง outbox; คืน `ManageProfile` ล่าสุด
      operationId: updateContactByHr
      security:
        - keycloak: [personnel:manage:person]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContactPatch' }
      responses:
        '200':
          description: แก้แล้ว (หรือไม่มีอะไรเปลี่ยน)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ManageProfile' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /persons/{personId}/emergency-contacts:
    parameters:
      - $ref: '#/components/parameters/PersonId'
    put:
      tags: [Provisioning]
      summary: Replace the emergency contacts (max 3)
      description: |
        แทนที่รายการผู้ติดต่อฉุกเฉินทั้งหมด (สูงสุด 3 คน, `priority` 1-3 ไม่ซ้ำ; รายการว่าง = ลบทั้งหมด)
        `data_change_log` บันทึกเฉพาะ "ฟิลด์/ช่องที่เปลี่ยน" ไม่เก็บค่า (ข้อมูลของบุคคลที่สาม) เหมือน `PUT /me/emergency-contacts`
        **สิทธิ์:** scope `personnel:manage:person` **และ** realm role `hr_master_data_admin` (ไม่มี role -> 403 `insufficient-role`)
        **ทุกการเขียน:** `expectedVersion` บังคับ (ไม่ตรง -> 409 `version-conflict`), `reason` บังคับ (ว่าง -> 422 `reason-required`, มีเลข 13 หลัก -> 422
        `reason-contains-pid`), บันทึก `data_change_log` พร้อมผู้กระทำ, บุคคลที่ยัง `PENDING_CLAIM` ไม่ส่ง outbox; คืน `ManageProfile` ล่าสุด
      operationId: replaceEmergencyContactsByHr
      security:
        - keycloak: [personnel:manage:person]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EmergencyContactsReplace' }
      responses:
        '200':
          description: แทนที่แล้ว (หรือไม่มีอะไรเปลี่ยน)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ManageProfile' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /persons/{personId}/history:
    parameters:
      - $ref: '#/components/parameters/PersonId'
    get:
      tags: [Provisioning]
      summary: Change history of this person for HR (newest first)
      description: |
        ประวัติการเปลี่ยนแปลงของบุคคลนี้ (ฟิลด์ใดเปลี่ยน ใคร เมื่อไหร่ เหตุผล) เรียงใหม่ -> เก่า ค่าของฟิลด์ชั้น CONFIDENTIAL/SENSITIVE/RESTRICTED ถูกปกปิด
        (`valuesHidden = true`) `actorSub` เป็นรหัสบัญชี (ยังไม่มีชื่อผู้กระทำ)
        **สิทธิ์:** scope `personnel:manage:person` **และ** realm role `hr_master_data_admin` (ตรวจที่ API ทุก request; ไม่มี role -> 403 `insufficient-role`)
        ทุกครั้งที่เปิดดูถูกบันทึกใน access log
      operationId: getPersonHistory
      security:
        - keycloak: [personnel:manage:person]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: ประวัติ
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/ChangeLogEntry' }
                  page: { $ref: '#/components/schemas/PageInfo' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # --------------------------------------------------------------------- Me
...
components:
  securitySchemes:
    keycloak:
      type: oauth2
      description: |
        Keycloak realm `lp-pao` เป็นผู้ออก token สำหรับเรียก MDM API (RS256, ตรวจสอบด้วย JWKS) โดยแยกจากเส้นทางล็อกอินของผู้ใช้
        ซึ่งยังคงเป็น thaid.lp-pao.go.th → check.lp-pao.go.th ตามเดิม
        - ระบบ ↔ ระบบ (รวม check.lp-pao.go.th และ backend ของทุกระบบปลายทาง) ใช้ **client credentials**
          และได้ scope ตาม client scope ที่ผู้ดูแล MDM/DPO อนุมัติเป็นราย client
        - ผู้ใช้ (self-service portal, HR/DPO console) ใช้ **authorization code + PKCE** กับ Keycloak
          ซึ่ง broker ไปยัง check.lp-pao.go.th ในฐานะ OIDC Identity Provider (ทางเลือก A ในเอกสารออกแบบ)
          และได้ scope ตาม realm role (`staff`, `hr_officer`, `dpo`)
        claim ที่ MDM ใช้: `sub`, `azp` (client id), `scope`, `person_id` (user context), `roles`
      flows:
        clientCredentials:
          tokenUrl: https://iam.lp-pao.go.th/realms/lp-pao/protocol/openid-connect/token
          scopes: &scopes
            personnel:read:basic: personId, ชื่อ-สกุล, ตำแหน่ง, สังกัด, ประเภทบุคลากร, สถานะ, version
            personnel:read:contact: เบอร์โทร, อีเมล, ที่อยู่ปัจจุบัน, ผู้ติดต่อฉุกเฉิน
            personnel:read:identity: วันเกิด, เพศ, ที่อยู่ตามทะเบียนบ้าน, วันออก/หมดอายุบัตร, สถานะ ThaID
            personnel:read:employment: วันบรรจุ, ระดับ, ประวัติการดำรงตำแหน่ง, เหตุพ้นจากราชการ
            personnel:read:photo: รูปถ่ายจาก ThaID
            personnel:read:inactive: เห็น record สถานะ INACTIVE (ลาออก/โอนย้าย)
            personnel:read:pid_masked: เห็นเลขบัตรประชาชนแบบปิด เหลือ 4 หลักท้าย (employeeNoMasked)
            personnel:read:pid: ถอดรหัสเลขบัตรประชาชน (อนุมัติโดย DPO เป็นราย client)
            personnel:lookup:pid: ค้นหา personId จาก pid (ใช้ตอน migrate/เชื่อมครั้งแรก)
            personnel:self: ข้อมูลของตนเอง และแก้ไขข้อมูลติดต่อ/ความยินยอม (user context เท่านั้น)
            personnel:write:employment: HR บันทึกตำแหน่ง/สังกัด/วันบรรจุ, ระงับ/คืนสถานะ, ขอ re-verify
            personnel:provision: HR เตรียม record ล่วงหน้าและจัดการ claim request
            personnel:import: นำเข้าข้อมูลชุดจากระบบ HR เดิม
            personnel:manage:reference: >-
              เพิ่ม/แก้ master data หน่วยงานและตำแหน่ง (T10) - ต้องมี realm role `hr_master_data_admin` ควบคู่เสมอ
              (API ตรวจทั้ง scope และ role เพราะ hr-console ผูก scope นี้ให้ทุกผู้ใช้ของ client)
            personnel:manage:person: >-
              HR จัดการข้อมูลบุคคลที่ไม่ใช่ของ ThaID (PR-D2: manage-profile, expected-identity, contact, emergency-contacts, history) - ต้องมี realm role
              `hr_master_data_admin` ควบคู่เสมอ (API ตรวจทั้ง scope และ role) ไม่ผูกกับฟิลด์ใดใน response ของ Person
            sync:thaid: check.lp-pao.go.th เท่านั้น (client `check-broker`)
            events:read: อ่านฟีดเหตุการณ์
            webhook:manage: จัดการ webhook subscription ของ client ตนเอง
            audit:read: อ่าน access log และ change log (DPO / ผู้ตรวจสอบ)
            audit:review: บันทึกผลรีวิวการเปิดเลขบัตร (เฉพาะ realm role dpo - MDM API ตรวจ role เพิ่มเสมอ ไม่ใช้ scope เป็นตัวกั้นเพียงอย่างเดียว)
        authorizationCode:
...
  parameters:
...
    PersonId:
      name: personId
      in: path
      required: true
      description: internal ID (UUID) ของบุคลากร
      schema: { type: string, format: uuid }
...
    Cursor:
      name: cursor
      in: query
      description: cursor จาก `page.nextCursor` ของหน้าก่อน
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
...
  responses:
    BadRequest:
      description: คำขอไม่ถูกต้อง
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    Forbidden:
      description: scope ไม่พอ หรือ client ถูกระงับ
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: ไม่พบ (หรือไม่มีสิทธิ์เห็น record นั้น)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: ข้อมูลขัดแย้ง (เช่น version ไม่ตรง, ตำแหน่งมีผู้ครองอยู่)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unprocessable:
      description: ข้อมูลถูกรูปแบบแต่ขัดกับ business rule (เช่น หน่วยงานที่อ้างถึงไม่มี/ไม่ active)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
...
    PersonStatus:
      type: string
      enum: [PENDING_CLAIM, ACTIVE, INACTIVE]
...
    Address:
      type: object
      properties:
        houseNo: { type: string }
        moo: { type: string }
        soi: { type: string }
        road: { type: string }
        subdistrict: { $ref: '#/components/schemas/AreaCode' }
        district: { $ref: '#/components/schemas/AreaCode' }
        province: { $ref: '#/components/schemas/AreaCode' }
        postcode: { type: string }
        fullText: { type: string, description: ที่อยู่แบบข้อความเต็ม }
    AreaCode:
      type: object
      properties:
        code: { type: string, description: รหัสตามกรมการปกครอง }
        nameTh: { type: string }
...
    ManageProfile:
      type: object
      description: |
        โปรไฟล์สำหรับฟอร์มแก้ไขของ HR (ไม่มี `x-required-scope` รายฟิลด์ - ควบคุมด้วย scope `personnel:manage:person` + role `hr_master_data_admin` ทั้ง endpoint)
        `expectedIdentity` คือข้อมูลที่ **HR กรอก รอยืนยันด้วย ThaID** ไม่ใช่ค่าจาก ThaID
      required: [personId, status, version, updatedAt, expectedIdentity, emergencyContacts]
      properties:
        personId: { type: string, format: uuid }
        status: { $ref: '#/components/schemas/PersonStatus' }
        version: { type: integer, description: ใช้เป็น `expectedVersion` ของการเขียนถัดไป }
        updatedAt: { type: string, format: date-time }
        expectedIdentity:
          type: object
          required: [editable, lockedReason]
          properties:
            firstNameTh: { type: string }
            lastNameTh: { type: string }
            birthDate: { type: string, format: date }
            editable: { type: boolean, description: แก้ได้เฉพาะคนที่ยังไม่เคยยืนยัน ThaID }
            lockedReason:
              type: [string, 'null']
              enum: [THAID_VERIFIED, NOT_PENDING_CLAIM, null]
        contact:
          type: object
          properties:
            mobilePhone: { type: [string, 'null'] }
            phoneAlt: { type: [string, 'null'] }
            emailPersonal: { type: [string, 'null'], format: email }
            lineId: { type: [string, 'null'] }
            sameAsRegistered: { type: boolean }
            currentAddress: { $ref: '#/components/schemas/Address' }
            updatedAt: { type: string, format: date-time }
            updatedBy: { type: string, enum: [SELF, HR] }
        emergencyContacts:
          type: array
          items: { $ref: '#/components/schemas/EmergencyContact' }
    ManageWriteBase:
      type: object
      required: [expectedVersion, reason]
      properties:
        expectedVersion: { type: integer, minimum: 1, description: version ที่ผู้เรียกเห็นล่าสุด (บังคับ) ไม่ตรง -> 409 `version-conflict` }
        reason: { type: string, minLength: 1, maxLength: 500, description: เหตุผล (บังคับ ห้ามมีเลข 13 หลัก) }
    ExpectedIdentityPatch:
      allOf:
        - $ref: '#/components/schemas/ManageWriteBase'
        - type: object
          description: ส่งอย่างน้อยหนึ่งฟิลด์
          properties:
            firstNameTh: { type: string, minLength: 1, maxLength: 200 }
            lastNameTh: { type: string, minLength: 1, maxLength: 200 }
            birthDate: { type: [string, 'null'], format: date, description: ช่วง 1900-01-01 ถึงวันนี้ (นอกช่วง -> 422); null = ล้างค่า }
    ContactPatch:
      allOf:
        - $ref: '#/components/schemas/ManageWriteBase'
        - type: object
          description: ส่งอย่างน้อยหนึ่งฟิลด์; ฟิลด์ที่ไม่ส่ง = ไม่แตะ, null/สตริงว่าง = ล้างค่า
          properties:
            mobilePhone: { type: [string, 'null'], maxLength: 20, pattern: '^(0[0-9]{8,9})?$' }
            phoneAlt: { type: [string, 'null'], maxLength: 20 }
            emailPersonal: { type: [string, 'null'], format: email, maxLength: 255 }
            lineId: { type: [string, 'null'], maxLength: 100 }
            sameAsRegistered: { type: boolean }
            currentAddress:
              type: object
              properties:
                houseNo: { type: [string, 'null'], maxLength: 50 }
                moo: { type: [string, 'null'], maxLength: 20 }
                soi: { type: [string, 'null'], maxLength: 100 }
                road: { type: [string, 'null'], maxLength: 100 }
                postcode: { type: [string, 'null'], maxLength: 10 }
                fullText: { type: [string, 'null'], maxLength: 2000 }
                subdistrict: { $ref: '#/components/schemas/AreaCodePatch' }
                district: { $ref: '#/components/schemas/AreaCodePatch' }
                province: { $ref: '#/components/schemas/AreaCodePatch' }
    AreaCodePatch:
      type: [object, 'null']
      properties:
        code: { type: string, maxLength: 10 }
    EmergencyContactsReplace:
      allOf:
        - $ref: '#/components/schemas/ManageWriteBase'
        - type: object
          required: [contacts]
          properties:
            contacts:
              type: array
              maxItems: 3
              items: { $ref: '#/components/schemas/EmergencyContact' }
    EmergencyContact:
      type: object
      required: [fullName, relationship, phone]
      properties:
        fullName: { type: string, minLength: 1, maxLength: 200 }
        relationship: { type: string, minLength: 1, maxLength: 50, examples: [คู่สมรส] }
        phone: { type: string, minLength: 1, maxLength: 20 }
        priority: { type: integer, minimum: 1, maximum: 3 }
...
    ChangeLogEntry:
      type: object
      properties:
        logId: { type: integer, format: int64 }
        changedAt: { type: string, format: date-time }
        fieldKey: { type: string, examples: [identity.reg_address_text] }
        oldValue: { description: ค่าเดิม (mask เมื่อฟิลด์เป็น RESTRICTED) }
        newValue: { description: ค่าใหม่ }
        changedBy: { type: string, enum: [THAID_SYNC, SELF, HR, HR_IMPORT, ADMIN] }
        actorSub: { type: [string, 'null'], description: "ผู้กระทำ: sub ของผู้ใช้, personId (SELF), หรือ system:thaid-sync / system:hr-import; null = แถวก่อน PR-A" }
        actorClient: { type: [string, 'null'], description: azp ของ token ที่เรียก }
        valuesHidden: { type: boolean, description: ค่าถูกปกปิดเพราะฟิลด์ชั้น CONFIDENTIAL/SENSITIVE/RESTRICTED }
        syncEventId: { type: [string, 'null'], format: uuid }
        reason: { type: [string, 'null'] }
...
    PageInfo:
      type: object
      properties:
        nextCursor: { type: [string, 'null'] }
        limit: { type: integer }
    Problem:
      type: object
      description: RFC 9457 Problem Details
      properties:
        type: { type: string, format: uri, examples: ['https://mdm.lp-pao.go.th/problems/insufficient-scope'] }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        requestId: { type: string }
        errors:
          type: array
          description: รายละเอียดรายฟิลด์สำหรับ 400
          items:
            type: object
            properties:
              field: { type: string }
              code: { type: string }
              message: { type: string }